refactor(*): refactor test files to be next what their testing
test files have been refactored to be next to the files their testing to use a make finding and
editing tests easier as well as to make it weasy to see if files have tests. also moved endpoint
middleware functions into their own files to increase modularity
