+++ .gitignore
@@ -3,3 +3,4 @@
 /node_modules/
 /.idea/
 /.nyc_output
+/coverage
