+++ api/hotMeals/hotMealRoutes.js
@@ -6,6 +6,7 @@ const socrataHotMealsAPI = config.get('resources.socrata.hot_meals');
 const addLatLngFromGoogle = require('./lib/addLatLngFromGoogle');
 const getHotMealLocations = require('./lib/getHotMealLocations');
 const getClosestHotMeal = require('./middleware/getClosestHotMeal');
+const getClosestHotMealEndpoint = require('./middleware/getClosestHotMealEndpoint');
 
 // get closest hot meal locations from API then repeat once every 24 hours
 getHotMealLocations(socrataHotMealsAPI.seattle, addLatLngFromGoogle);
@@ -13,6 +14,6 @@ setInterval(() => {
   getHotMealLocations(socrataHotMealsAPI, addLatLngFromGoogle);
 }, 86400000);
 
-router.get('/hot-meal/closest', getClosestHotMeal.get, getClosestHotMeal.endpoint);
+router.get('/hot-meal/closest', getClosestHotMeal, getClosestHotMealEndpoint);
 
 module.exports = exports = router;
