+++ api/hotMeals/middleware/getClosestHotMeal.js
@@ -35,15 +35,5 @@ function getClosestHotMeal(req, res, next) {
   }
 }
 
-function getClosestHotMealEndPoint(req, res) {
-  const status = req.returnVal && req.returnVal.status ? req.returnVal.status : 400;
-  const data = req.returnVal && req.returnVal.data ? req.returnVal.data : { 'error' : 'sorry we couldn\'t interpret you\'re request' };
-  res.status(status);
-  res.json(data);
 
-}
-
-module.exports = exports = {
-  get: getClosestHotMeal,
-  endpoint: getClosestHotMealEndPoint
-};
+module.exports = exports = getClosestHotMeal;
