+++ api/hotMeals/middleware/getClosestHotMealEndpoint.js
@@ -1,0 +1,11 @@
+'use strict';
+
+function getClosestHotMealEndPoint(req, res) {
+  const status = req.returnVal && req.returnVal.status ? req.returnVal.status : 400;
+  const data = req.returnVal && req.returnVal.data ? req.returnVal.data : { 'error' : 'sorry we couldn\'t interpret you\'re request' };
+  res.status(status);
+  res.json(data);
+
+}
+
+module.exports = exports = getClosestHotMealEndPoint;
