+++ lib/errorHandler/errorHandler.test.js
@@ -23,7 +23,7 @@ describe('errorHandler', function () {
         expect(error).to.deep.equal(testError);
       }
     };
-    let errorHandler = proxyquire('../errorHandler', {'winston': winstonStub});
+    let errorHandler = proxyquire('./errorHandler', {'winston': winstonStub});
     const nextSpy = sinon.spy();
     const res = {
       status: sinon.spy(),
@@ -48,7 +48,7 @@ describe('errorHandler', function () {
         expect(error).to.deep.equal(testError);
       }
     };
-    let errorHandler = proxyquire('../errorHandler', {'winston': winstonStub});
+    let errorHandler = proxyquire('./errorHandler', {'winston': winstonStub});
     const nextSpy = sinon.spy();
     const res = {
       status: sinon.spy(),
