+++ lib/getDistanceFromLatLng/getDistanceFromLatLng.test.js
@@ -2,7 +2,7 @@
 
 const expect = require('chai').expect;
 
-const getDistanceFromLatLng = require('../getDistanceFromLatLng');
+const getDistanceFromLatLng = require('./getDistanceFromLatLng');
 
 describe('getDistanceFromLatLng', function () {
